phonemesync/backend/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
aiofiles==23.2.1

# Data validation
pydantic==2.7.1
pydantic-settings==2.2.1

# Job store
redis==5.0.4
rq==1.16.1

# ML — Wav2Lip dependencies
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2
numpy==1.26.4
opencv-python-headless==4.9.0.80
scipy==1.13.0
librosa==0.10.2
numba==0.59.1
Pillow==10.3.0
tqdm==4.66.4
face-alignment==1.4.1
batch-face==1.5.3

# Whisper
openai-whisper==20231117
tiktoken==0.6.0

# Phoneme mapping
pronouncing==0.2.0
nltk==3.8.1

# MediaPipe
mediapipe==0.10.13

# Audio processing
soundfile==0.12.1
pydub==0.25.1

# Video processing (FFmpeg python wrapper)
ffmpeg-python==0.2.0

# Utilities
python-dotenv==1.0.1
structlog==24.1.0
tenacity==8.2.3
httpx==0.27.0

# Testing
pytest==8.2.0
pytest-asyncio==0.23.7
